requirements.txt
aiohttp==3.9.3
aiogram==3.4.1
python-dotenv==1.0.1
